feat: process primarily

+++ src/main.cpp
@@ -1,5 +1,6 @@
 #include <args/CmdArgs.h>
-#include <filesystem>
+#include <cstdio>
+#include <format.h>
 #include <iostream>
 
 /*
@@ -14,26 +15,65 @@ Optional arguments:
 
 namespace fs = std::filesystem;
 
-int main(int argc, char** argv) {
-    auto [inputPath, outputPath] = parseCmdArgs(argc, argv);
-
-    if (!fs::is_directory(inputPath) || !fs::is_directory(outputPath)) {
-        std::cerr << "Input or output path is not a directory" << std::endl;
-        return 1;
-    }
-
+bool checkPath(fs::path const& inputPath, fs::path const& outputPath) {
     if (!fs::exists(inputPath)) {
         std::cerr << "Input path does not exist" << std::endl;
-        return 1;
+        return false;
     }
 
     if (!fs::exists(outputPath)) {
         fs::create_directory(outputPath);
     }
 
+    if (!fs::is_directory(inputPath) || !fs::is_directory(outputPath)) {
+        std::cerr << "Input or output path is not a directory" << std::endl;
+        return false;
+    }
+
     if (!fs::is_empty(outputPath)) {
-        std::cout << "Output directory is not empty, files will be overwritten." << std::endl;
+        std::cout << "Output directory is not empty, files will be overwritten [Y/n]: ";
+        std::string input;
+        std::getline(std::cin, input, '\n');
+        if (input == "n" || input == "N") {
+            exit(0);
+        }
         fs::remove_all(outputPath);
+    }
+
+    return true;
+}
+
+// def check_custom_data_dir(output_arg) -> str:
+//     if os.path.basename(output_arg) != "testdata":
+//         output_arg = os.path.join(output_arg, "testdata")
+//         logger.warning(f"Output directory should be named as testdata. The data will convert to {output_arg}.")
+//     return output_arg
+
+fs::path checkCustomDataDir(fs::path const& outputPath) {
+    if (outputPath.filename() != "testdata") {
+        auto newOutputPath = outputPath / "testdata";
+        std::cerr << "Output directory should be named as testdata. The data will convert to "
+                  << newOutputPath.string() << std::endl;
+        return newOutputPath;
+    }
+    return outputPath;
+}
+
+int main(int argc, char** argv) {
+    auto const& [inputPath, outputPath] = parseCmdArgs(argc, argv);
+
+    if (!checkPath(inputPath, outputPath))
+        return 1;
+
+    std::cout << "Start to convert the data\nInput directory: " << fs::absolute(inputPath).string()
+              << "\nOutput directory: " << fs::absolute(outputPath).string() << std::endl;
+
+    if (isCustomData(inputPath)) {
+    } else if (isHydroExport(inputPath)) {
+    } else {
+        std::cerr << "Input path is not a custom data directory or a hydro export directory"
+                  << std::endl;
+        return 1;
     }
 
     return 0;
